Improved types part 2

+++ Engine/include/Engine/PhysicsHandler.h
@@ -8,9 +8,9 @@ namespace engine {
 
 /// @brief Contains collision detection result data.
 struct collisionHit {
-    utils::f2v point;  ///< Contact point of collision.
-    utils::f2v normal; ///< Surface normal at collision point.
-    float distance;    ///< Distance to collision point.
+    utils::f2v point;   ///< Contact point of collision.
+    utils::f2v normal;  ///< Surface normal at collision point.
+    float32_t distance; ///< Distance to collision point.
 };
 
 class PhysicsHandler : public ISettingsObserver {
@@ -19,17 +19,17 @@ public:
     ~PhysicsHandler();
 
     /// @brief Applies gravity force to entity based on ground state and delta time.
-    void applyGravity(DynamicEntity&, float deltaTime) const;
+    void applyGravity(DynamicEntity&, float32_t deltaTime) const;
 
-    void setGAcceleration(float g) { _gAcceleration = g; }
+    void setGAcceleration(float32_t g) { _gAcceleration = g; }
 
     /// @brief Handles collisions between entity and map tiles.
-    void handleMapCollisions(DynamicEntity&, const TileLayer::Grid&, float deltaTime) const;
+    void handleMapCollisions(DynamicEntity&, const TileLayer::Grid&, float32_t deltaTime) const;
 
     /// @brief Handles attack collisions between player and an entity.
     void handleAttacksCollisions(DynamicSpriteEntity& player, DynamicEntity& entity) const;
     
-    void setFallGMultiplier(float fallGMultiplier) { _fallGMultiplier = fallGMultiplier; }
+    void setFallGMultiplier(float32_t fallGMultiplier) { _fallGMultiplier = fallGMultiplier; }
 
     void onSettingsChanged() override;
 
@@ -40,19 +40,19 @@ private:
         const SDL_Rect& target,
         const utils::f2v& velocity,
         collisionHit& outhit,
-        float deltaTime) const;
+        float32_t deltaTime) const;
 
     /// @brief Checks for horizontal collisions with map tiles.
     void checkHorizontalCollision(
         const TileLayer::Grid&, 
-        float deltaTime,
+        float32_t deltaTime,
         utils::f2v& velocity,
         const SDL_FRect& hitBox) const;
     
     /// @brief Checks for vertical collisions with map tiles.
     void checkVerticalCollision(
         const TileLayer::Grid&,
-        float deltaTime,
+        float32_t deltaTime,
         utils::f2v& velocity,
         const SDL_FRect& hitBox,
         bool& landed) const;
@@ -61,15 +61,15 @@ private:
     void checkIfStillOnGround(
         DynamicEntity& entity,
         const TileLayer::Grid&,
-        float deltaTime,
+        float32_t deltaTime,
         const SDL_FRect& hitBox,
         const utils::f2v& velocity) const;
 
-    Settings& _settings;      ///< Reference to game settings.
+    Settings& _settings;          ///< Reference to game settings.
 
-    size_t _tileSize{};       ///< Size of map tiles in pixels.
-    float _gAcceleration{};   ///< Gravitational acceleration value.
-    float _fallGMultiplier{}; ///< Multiplier for faster falling.
+    uint32_t _tileSize{};         ///< Size of map tiles in pixels.
+    float32_t _gAcceleration{};   ///< Gravitational acceleration value.
+    float32_t _fallGMultiplier{}; ///< Multiplier for faster falling.
 };
 
 }
